fix: cast make_interval parameter to integer for login lockout check
make_interval(mins => bigint) does not exist in PostgreSQL — the
named parameter requires integer type. Previously hidden by
unwrap_or((0,)), exposed after switching to ? error propagation.

+++ backend/src/services/auth/login.rs
@@ -18,7 +18,7 @@ impl AuthService {
                 SELECT COUNT(*) FROM login_events
                 WHERE email = $1
                   AND event_type = 'login_failure'
-                  AND created_at > NOW() - make_interval(mins => $2)
+                  AND created_at > NOW() - make_interval(mins => $2::integer)
                 "#,
             )
             .bind(&req.email)
